Множественный импорт сетей, доработка задач
Разработан маршрут для множественного импорта сетей.
Для project_helpers частично сгенерирована документация.
При добавлении файла к задаче - он автоматически добавляется в webfiles в директорию "Tasks files" (в зависимости от локализации).

+++ app/helpers/main_page_helpers.py
@@ -80,31 +80,43 @@ class DefaultSidebar:
 
 class DefaultEnvironment:
     def __init__(self, obj: str, op: str, **kwargs):
+        from app.helpers.roles import user_position_can_make_action
+        acts = []
         # Title page
         if obj == 'main_page' and op == 'show':
             current_object = CurrentObjectInfo(current_app.config["GlobalSettings"].main_page_name, "fa-solid fa-user-secret")
             title = _l("Title page")
         # Projects
         elif obj == 'Project' and op == 'index':
-            act1 = CurrentObjectAction(_l("Add new project"), "fa-solid fa-square-plus", url_for('projects.project_new'))
-            current_object = CurrentObjectInfo(_l("List of all projects"), "fa-solid fa-list-check", subtitle=_l('Projects, created in PCF'), actions=[act1])
+            if user_position_can_make_action(current_user, models.Project, 'create'):
+                act1 = CurrentObjectAction(_l("Add new project"), "fa-solid fa-square-plus", url_for('projects.project_new'))
+                acts.append(act1)
+            current_object = CurrentObjectInfo(_l("List of all projects"), "fa-solid fa-list-check", subtitle=_l('Projects, created in PCF'), actions=acts)
             title = _l("All projects")
         elif obj == "Project" and op == 'new':
             current_object = CurrentObjectInfo(_l("Add new project"), "fa-solid fa-square-plus")
             title = _l("Add new project")
         # Wiki pages
         elif obj == 'WikiPage' and op == 'index':
-            act1 = CurrentObjectAction(_l("Add new page"), "fa-solid fa-file-circle-plus", url_for('wiki_pages.wikipage_new'))
-            act2 = CurrentObjectAction(_l("Add new directory"), 'fa-solid fa-folder-plus', url_for('wiki_pages.pagedirectory_new'))
-            current_object = CurrentObjectInfo(_l('Wiki pages'), "fa-brands fa-wikipedia-w", actions=[act1, act2])
+            if user_position_can_make_action(current_user, models.WikiPage, 'create'):
+                act1 = CurrentObjectAction(_l("Add new page"), "fa-solid fa-file-circle-plus", url_for('wiki_pages.wikipage_new'))
+                acts.append(act1)
+            if user_position_can_make_action(current_user, models.WikiDirectory, 'create'):
+                act2 = CurrentObjectAction(_l("Add new directory"), 'fa-solid fa-folder-plus', url_for('wiki_pages.pagedirectory_new'))
+                acts.append(act2)
+            current_object = CurrentObjectInfo(_l('Wiki pages'), "fa-brands fa-wikipedia-w", actions=acts)
             title = _l("Knowledge base")
         elif obj == 'WikiPage' and op == 'new':
             current_object = CurrentObjectInfo(_l("Add new Wiki page"), 'fa-brands fa-wikipedia-w')
             title = _l("Add a page to the knowledge base")
         elif obj == 'WikiPage' and op == 'show':
-            act1 = CurrentObjectAction(_l("Edit"), "fa-solid fa-square-pen", url_for('wiki_pages.wikipage_edit', wikipage_id=kwargs['obj_val'].id))
-            act2 = CurrentObjectAction(_l("Delete"), "fa-solid fa-trash", url_for('wiki_pages.wikipage_delete', wikipage_id=kwargs['obj_val'].id), confirm=_l("Are you sure you want to delete this page?"), btn_class='btn-danger', method='DELETE')
-            current_object = CurrentObjectInfo(_l("Knowledge Base Page"), 'fa-brands fa-wikipedia-w', actions=[act1, act2])
+            if user_position_can_make_action(current_user, models.WikiPage, 'update'):
+                act1 = CurrentObjectAction(_l("Edit"), "fa-solid fa-square-pen", url_for('wiki_pages.wikipage_edit', wikipage_id=kwargs['obj_val'].id))
+                acts.append(act1)
+            if user_position_can_make_action(current_user, models.WikiPage, 'delete'):
+                act2 = CurrentObjectAction(_l("Delete"), "fa-solid fa-trash", url_for('wiki_pages.wikipage_delete', wikipage_id=kwargs['obj_val'].id), confirm=_l("Are you sure you want to delete this page?"), btn_class='btn-danger', method='DELETE')
+                acts.append(act2)
+            current_object = CurrentObjectInfo(_l("Knowledge Base Page"), 'fa-brands fa-wikipedia-w', actions=acts)
             title = kwargs['obj_val'].title
         elif obj == 'WikiPage' and op == 'edit':
             current_object = CurrentObjectInfo(_l("Editing the knowledge base page"), 'fa-brands fa-wikipedia-w')
@@ -114,16 +126,22 @@ class DefaultEnvironment:
             title = _l("Add new directory to knowledge base")
         # Critical vulnerabilities
         elif obj == 'CriticalVulnerability' and op == 'index':
-            act1 = CurrentObjectAction(_l("Add new critical vulnerability"), 'fa-solid fa-square-plus', url_for('cves.cve_new'))
-            current_object = CurrentObjectInfo(_l("CVE list"), 'fa-solid fa-droplet', actions=[act1])
+            if user_position_can_make_action(current_user, models.CriticalVulnerability, 'create'):
+                act1 = CurrentObjectAction(_l("Add new critical vulnerability"), 'fa-solid fa-square-plus', url_for('cves.cve_new'))
+                acts.append(act1)
+            current_object = CurrentObjectInfo(_l("CVE list"), 'fa-solid fa-droplet', actions=acts)
             title = _l("Critical vulnerabilities")
         elif obj == "CriticalVulnerability" and op == "new":
             current_object = CurrentObjectInfo(_l("Add a new entry about a critical vulnerability"), "fa-solid fa-square-plus")
             title = _l("Add new critical vulnerability")
         elif obj == "CriticalVulnerability" and op == "show":
-            act1 = CurrentObjectAction(_l("Edit"), "fa-solid fa-square-pen", url_for('cves.cve_edit', cve_id=kwargs['obj_val'].id))
-            act2 = CurrentObjectAction(_l("Delete"), "fa-solid fa-trash", url_for("cves.cve_delete", cve_id=kwargs['obj_val'].id, confirm=_l("Are you sure you want to delete this Critical Vulnerability?")), btn_class='btn-danger', method='DELETE')
-            current_object = CurrentObjectInfo(_l("Critical vulnerability «%(title)s»", title=kwargs['obj_val'].title), "fa-solid fa-droplet", actions=[act1, act2])
+            if user_position_can_make_action(current_user, models.CriticalVulnerability, 'update'):
+                act1 = CurrentObjectAction(_l("Edit"), "fa-solid fa-square-pen", url_for('cves.cve_edit', cve_id=kwargs['obj_val'].id))
+                acts.append(act1)
+            if user_position_can_make_action(current_user, models.CriticalVulnerability, 'delete'):
+                act2 = CurrentObjectAction(_l("Delete"), "fa-solid fa-trash", url_for("cves.cve_delete", cve_id=kwargs['obj_val'].id, confirm=_l("Are you sure you want to delete this Critical Vulnerability?")), btn_class='btn-danger', method='DELETE')
+                acts.append(act2)
+            current_object = CurrentObjectInfo(_l("Critical vulnerability «%(title)s»", title=kwargs['obj_val'].title), "fa-solid fa-droplet", actions=acts)
             title = _l("Critical vulnerability «CVE-%(year)s-%(identifier)s»", year=kwargs['obj_val'].year, identifier=kwargs['obj_val'].identifier)
         elif obj == 'CriticalVulnerability' and op == 'edit':
             current_object = CurrentObjectInfo(_l("Edit critical vulnerability"), "fa-solid fa-square-pen")
@@ -144,7 +162,7 @@ class DefaultEnvironment:
         elif obj == 'Team' and op == 'team_index':
             title = _l("All teams")
             acts = []
-            if not current_user.is_anonymous and current_user.position.is_administrator:
+            if user_position_can_make_action(current_user, models.Team, 'create'):
                 act1 = CurrentObjectAction(_l("Add new team"), "fa-solid fa-square-plus", url_for('users.team_new'))
                 acts.append(act1)
             current_object = CurrentObjectInfo(_l("Team list"), models.Team.Meta.icon, subtitle=models.Team.Meta.description, actions=acts)
@@ -153,9 +171,13 @@ class DefaultEnvironment:
             current_object = CurrentObjectInfo(title, "fa-solid fa-square-plus", subtitle=models.Team.Meta.description)
         elif obj == 'Team' and op == 'team_show':
             title = _l("Team #%(team_id)s", team_id=kwargs['obj_val'].id)
-            act1 = CurrentObjectAction(_l("Edit"), "fa-solid fa-square-pen", url_for('users.team_edit', team_id=kwargs['obj_val'].id))
-            act2 = CurrentObjectAction(_l("Delete"), "fa-solid fa-trash", url_for('users.team_delete', team_id=kwargs['obj_val'].id), confirm=_l("Are you sure you want to delete this team?"), btn_class='btn-danger', method='DELETE')
-            current_object = CurrentObjectInfo(_l("Team #%(team_id)s: «%(team_title)s»", team_id=kwargs['obj_val'].id, team_title=kwargs['obj_val'].title), models.Team.Meta.icon, subtitle=models.Team.Meta.description, actions=[act1, act2])
+            if user_position_can_make_action(current_user, models.Team, 'update'):
+                act1 = CurrentObjectAction(_l("Edit"), "fa-solid fa-square-pen", url_for('users.team_edit', team_id=kwargs['obj_val'].id))
+                acts.append(act1)
+            if user_position_can_make_action(current_user, models.Team, 'delete'):
+                act2 = CurrentObjectAction(_l("Delete"), "fa-solid fa-trash", url_for('users.team_delete', team_id=kwargs['obj_val'].id), confirm=_l("Are you sure you want to delete this team?"), btn_class='btn-danger', method='DELETE')
+                acts.append(act2)
+            current_object = CurrentObjectInfo(_l("Team #%(team_id)s: «%(team_title)s»", team_id=kwargs['obj_val'].id, team_title=kwargs['obj_val'].title), models.Team.Meta.icon, subtitle=models.Team.Meta.description, actions=acts)
         elif obj == 'Team' and op == 'team_edit':
             title = _l("Edit team #%(team_id)s", team_id=kwargs['obj_val'].id)
             current_object = CurrentObjectInfo(title, "fa-solid fa-square-pen")
